refactoring and design adjustments
