+++ frontend/language/src/nb.json
@@ -1467,11 +1467,11 @@
   "ux_editor.component_properties.summary.override.display_type.list": "Liste",
   "ux_editor.component_properties.summary.override.display_type.not_set": "Ikke satt",
   "ux_editor.component_properties.summary.override.display_type.string": "Tekst",
-  "ux_editor.component_properties.summary.override.empty_field_text": "Tekst du vil vise i tomt felt",
-  "ux_editor.component_properties.summary.override.force_show": "Vis alltid feltet",
+  "ux_editor.component_properties.summary.override.empty_field_text": "Tekst du vil vise i tomme felt",
+  "ux_editor.component_properties.summary.override.force_show": "Vis komponenten",
   "ux_editor.component_properties.summary.override.hidden": "Skjul feltet",
-  "ux_editor.component_properties.summary.override.hide_empty_fields": "Vis tomme felter",
-  "ux_editor.component_properties.summary.override.hide_empty_fields.info_message": "Du kan bare skjule tomme felt i oppsummeringen hvis de er valgfrie",
+  "ux_editor.component_properties.summary.override.hide_empty_fields": "Vis tomme felt",
+  "ux_editor.component_properties.summary.override.hide_empty_fields.info_message": "Du kan bare skjule komponenten fra oppsummeringen hvis feltet er valgfritt.",
   "ux_editor.component_properties.summary.override.is_compact": "Bruk kompakt visning",
   "ux_editor.component_properties.summary.override.show_component": "Vis komponenten",
   "ux_editor.component_properties.summaryDelimiter": "Skillelinje for sammendragsvisningsceller",
